git-svn-id: svn://svn.r-forge.r-project.org/svnroot/blockmodeling@91 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/blockmodelingTest/src/REGE_NM_DIAG_R.f90
@@ -62,7 +62,7 @@
       CM=INT(CM + b (max (i,j), min (i,j))*(min(R(I,I,1),r(j,j,1))+min(R(I,I,2),r(j,j,2))))
   505  CONTINUE
 !     COMPUTE REGULAR EQUIVALENCE
-  506 DM = DEG(II)+DEG(JJ)
+  506 DM = INT(DEG(II))+INT(DEG(JJ))
       B (II,JJ)= 1.0
       IF(DM.NE.0.0) B (II,JJ)=CM/DM
 !     IF(B (II,JJ).LE.CUT) B (II,JJ)=0.0
